update pipeline sdk 2e sample
