+++ sdk/jobs/pipelines/2e_image_classification_keras_minist_convnet/prep/prep_dsl_component.py
@@ -1,3 +1,5 @@
+# Converts MNIST-formatted files at the passed-in input path to training data output path and test data output path
+import os
 from pathlib import Path
 from azure.ml import dsl, Input, Output
 from azure.ml.entities import Environment
@@ -15,12 +17,42 @@ conda_env = Environment(
     description="Convert data to CSV file, and split to training and test data",
     environment=conda_env,
 )
-def prep(
+def prepare_data_component(
     input_data: Input,
     training_data: Output,
     test_data: Output,
 ):
-    # Avoid dependency issue, execution logic is in prep.py file
-    from prep import prep
+    convert(
+        os.path.join(input_data, "train-images-idx3-ubyte"),
+        os.path.join(input_data, "train-labels-idx1-ubyte"),
+        os.path.join(training_data, "mnist_train.csv"),
+        60000,
+    )
+    convert(
+        os.path.join(input_data, "t10k-images-idx3-ubyte"),
+        os.path.join(input_data, "t10k-labels-idx1-ubyte"),
+        os.path.join(test_data, "mnist_test.csv"),
+        10000,
+    )
 
-    prep(input_data, training_data, test_data)
+
+def convert(imgf, labelf, outf, n):
+    f = open(imgf, "rb")
+    l = open(labelf, "rb")
+    o = open(outf, "w")
+
+    f.read(16)
+    l.read(8)
+    images = []
+
+    for i in range(n):
+        image = [ord(l.read(1))]
+        for j in range(28 * 28):
+            image.append(ord(f.read(1)))
+        images.append(image)
+
+    for image in images:
+        o.write(",".join(str(pix) for pix in image) + "\n")
+    f.close()
+    o.close()
+    l.close()
