+++ sdk/jobs/pipelines/2e_image_classification_keras_minist_convnet/score/score.py
@@ -13,7 +13,6 @@ import numpy as np
 import pandas as pd
 import os
 import matplotlib.pyplot as plt
-from sklearn.model_selection import train_test_split
 import mlflow
